requirements.txt
# ===================================================
# متطلبات مشروع TurkTextileHub
# بوت تيليغرام للتجار الأتراك - تكامل KAYISOFT API
# ===================================================

# بوت تيليغرام
python-telegram-bot==20.7

# متغيرات البيئة
python-dotenv==1.0.0

# طلبات HTTP (Supabase REST API + KAYISOFT API + MinIO)
requests==2.31.0

# الذكاء الاصطناعي - استخراج بيانات المنتجات
openai>=1.0.0
